TUO 2.2.1 with batch files
TUO 2.2.1 with batch files  Just extract.  Cards and mission are updated
as of 25 Oct 2014.  If you need to update your cards and mission double
click on update.bat and choose yes to let powershell update files.

You will need to make a owncards.txt for each person you want to run in
the data folder.  Then open the members.txt file and add one member per
line as it is in the downloaded included file.   You can have 1 - 200
members.